Move `Shape` to dedicated module
I'm working on adding more infrastructure around `Shape`, to eliminate
any internal redundancy. It makes sense to have a dedicated module for
all that stuff.
