+++ src/kernel/mod.rs
@@ -78,5 +78,6 @@
 
 pub mod algorithms;
 pub mod geometry;
+pub mod shape;
 pub mod shapes;
 pub mod topology;
